requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.2.0
altair==5.0.1
anyio==3.7.0
appnope==0.1.3
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.12.2
bitarray==1.2.2
black==22.3.0
bleach==6.0.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2021.10.8
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.0.12
chromadb==0.3.26
click==8.1.6
clickhouse-connect==0.6.6
coloredlogs==15.0.1
comm==0.1.3
cryptography==41.0.3
cytoolz==0.11.2
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.1
dm-tree==0.1.8
duckdb==0.8.1
duckduckgo-search==3.8.4
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
exceptiongroup==1.1.1
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.100.0
fastjsonschema==2.17.1
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.3.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-search-results==2.4.2
grpcio==1.54.2
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hexbytes==0.2.2
hnswlib @ git+https://github.com/nmslib/hnswlib.git@0df757e5e3bf8c85300cc19d29d5c8098a206151
hpack==4.0.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.3
importlib-metadata==6.6.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.12.0
langchain==0.0.260
langchainplus-sdk==0.0.20
langsmith==0.0.20
libclang==16.0.0
llama-cpp-python==0.1.72
lru-dict==1.1.7
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.1.0
monotonic==1.6
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
netaddr==0.8.0
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
overrides==7.3.1
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipx==1.0.0
platformdirs==2.5.2
pluggy==1.0.0
posthog==3.0.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
py==1.11.0
py-solc-x==1.1.1
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.8
pypdf==3.13.0
PyPDF2==3.0.1
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.7
python-magic==0.4.27
python-version==0.0.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rlp==2.0.1
rsa==4.9
scipy==1.10.1
semantic-version==2.9.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
streamlit==1.24.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
termcolor==2.3.0
terminado==0.17.1
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.9.0
urllib3==1.26.9
userpath==1.8.0
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
varint==1.0.2
watchfiles==0.19.0
wcwidth==0.2.6
web3==5.29.0
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.3.4
widgetsnbextension==4.0.7
wikipedia==1.4.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.15.0
zstandard==0.21.0
